Did serveral things:  - Reorganized the code. This makes the code more maintanable  - Added exception handling on points when opening and writing files
